requirements.txt

apache-airflow==2.6.2
dbt-core==1.5.0
dbt-postgres==1.5.0
pandas==2.1.1
pyarrow==13.0.0
requests==2.31.0
flake8==6.1.0
docker==6.1.3
